pyproject.toml
[project]
name = "django-acquisitions"
version = "0.1.0"
description = "Django customer acquisition pipeline for managing leads, outreach, and onboarding"
readme = "README.md"
license = "BSD-3-Clause"
authors = [
    { name = "specialorange" },
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/specialorange/django-acquisitions"
Repository = "https://github.com/specialorange/django-acquisitions"

[project.optional-dependencies]
celery = ["celery>=5.0"]
twilio = ["twilio>=8.0"]
drf = ["djangorestframework>=3.14"]
all = [
    "celery>=5.0",
    "twilio>=8.0",
    "djangorestframework>=3.14",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "pre-commit>=3.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["acquisitions"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["acquisitions"]
